+++ uip/mhparse.c
@@ -2684,7 +2684,7 @@ invalid_digest:
 	    while (*cp)
 		cp++;
 	    fprintf (stderr, "invalid MD5 digest (got %d octets)\n",
-		     cp - bp);
+		     (int)(cp - bp));
 	}
 
 	return NOTOK;
